test(cli): split search flow coverage
